Move LibraryHeaderHolder to a new file
